Feat/integrate vnpay (#9)
* added:service and impl Booking service

* added:Booking repository

* added: service ,impl and repos Payment entity

* changed: Long to Integer of Id property

* added: config VNPayConfig

* added: VNPayService service

* added: UrlUtils and VNPayController controller

+++ src/main/java/com/trainticketbooking/app/Entities/Payment.java
@@ -12,7 +12,7 @@ public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "payment_id")
-    private Long paymentId;
+    private Integer paymentId;
 
     @OneToOne
     @JoinColumn(name = "booking_id", nullable = false)
